+++ test/unit/test_noise_learner_result.py
@@ -22,7 +22,7 @@ from qiskit_aer import AerSimulator
 
 from qiskit_ibm_runtime.fake_provider.local_service import QiskitRuntimeLocalService
 from qiskit_ibm_runtime.utils.noise_learner_result import PauliLindbladError, LayerError
-from qiskit_ibm_runtime.utils.visualization_utils import get_qubits_coordinates
+from qiskit_ibm_runtime.visualization.utils import get_qubits_coordinates
 
 from ..ibm_test_case import IBMTestCase
